+++ dolen_ig4_1_common.h
@@ -57,8 +57,8 @@ typedef struct {
     float *_v;
     float *_att;
     float *_logits;
-    float *_key_cache;
-    float *_value_cache;
+    _Float16 *_key_cache;
+    _Float16 *_value_cache;
     qtensor xq;
     qtensor hq;
     float *_cos_cache;
